+++ packages/atproto/lib/src/service/entities/record.dart
@@ -3,6 +3,7 @@
 // modification, are permitted provided the conditions.
 
 // 📦 Package imports:
+import 'package:atproto_core/atproto_core.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'record.freezed.dart';
@@ -12,7 +13,7 @@ part 'record.g.dart';
 class Record with _$Record {
   const factory Record({
     required String cid,
-    required String uri,
+    @AtUriConverter() required AtUri uri,
   }) = _Record;
 
   factory Record.fromJson(Map<String, Object?> json) => _$RecordFromJson(json);