+++ packages/atproto/lib/src/service/entities/record.g.dart
@@ -14,7 +14,8 @@ _$_Record _$$_RecordFromJson(Map json) => $checkedCreate(
       ($checkedConvert) {
         final val = _$_Record(
           cid: $checkedConvert('cid', (v) => v as String),
-          uri: $checkedConvert('uri', (v) => v as String),
+          uri: $checkedConvert(
+              'uri', (v) => const AtUriConverter().fromJson(v as String)),
         );
         return val;
       },
@@ -22,5 +23,5 @@ _$_Record _$$_RecordFromJson(Map json) => $checkedCreate(
 
 Map<String, dynamic> _$$_RecordToJson(_$_Record instance) => <String, dynamic>{
       'cid': instance.cid,
-      'uri': instance.uri,
+      'uri': const AtUriConverter().toJson(instance.uri),
     };