+++ packages/atproto_core/lib/atproto_core.dart
@@ -2,12 +2,14 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
+export 'package:at_uri/at_uri.dart' show AtUri;
 export 'package:atproto_core/src/base_service.dart';
 export 'package:atproto_core/src/client/challenge.dart';
 export 'package:atproto_core/src/client/client_context.dart';
 export 'package:atproto_core/src/client/jitter.dart';
 export 'package:atproto_core/src/client/retry_event.dart';
 export 'package:atproto_core/src/config/retry_config.dart';
+export 'package:atproto_core/src/converter/at_uri_converter.dart';
 export 'package:xrpc/xrpc.dart'
     show
         XRPCResponse,