+++ packages/bluesky/lib/src/service/entities/embed_record.g.dart
@@ -14,7 +14,8 @@ _$_EmbedRecord _$$_EmbedRecordFromJson(Map json) => $checkedCreate(
       ($checkedConvert) {
         final val = _$_EmbedRecord(
           cid: $checkedConvert('cid', (v) => v as String),
-          uri: $checkedConvert('uri', (v) => v as String),
+          uri: $checkedConvert(
+              'uri', (v) => const AtUriConverter().fromJson(v as String)),
           author: $checkedConvert('author',
               (v) => Actor.fromJson(Map<String, Object?>.from(v as Map))),
           record: $checkedConvert('record',
@@ -27,7 +28,7 @@ _$_EmbedRecord _$$_EmbedRecordFromJson(Map json) => $checkedCreate(
 Map<String, dynamic> _$$_EmbedRecordToJson(_$_EmbedRecord instance) =>
     <String, dynamic>{
       'cid': instance.cid,
-      'uri': instance.uri,
+      'uri': const AtUriConverter().toJson(instance.uri),
       'author': instance.author.toJson(),
       'record': instance.record.toJson(),
     };