+++ packages/bluesky/lib/src/service/entities/reposted_by_data.dart
@@ -3,6 +3,7 @@
 // modification, are permitted provided the conditions.
 
 // 📦 Package imports:
+import 'package:atproto_core/atproto_core.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 // 🌎 Project imports:
@@ -15,7 +16,7 @@ part 'reposted_by_data.g.dart';
 class RepostedByData with _$RepostedByData {
   const factory RepostedByData({
     required List<Actor> repostedBy,
-    required String uri,
+    @AtUriConverter() required AtUri uri,
     required String cursor,
   }) = _RepostedByData;